pyproject.toml
[project]
name = "leaves"
version = "0.1.0"
description = "Leaves - Connect sensors & actuators to the cloud"
authors = [{ name = "Bernhard Boser", email = "bernhard.e.boser@gmail.com" }]
readme = "README.md"
requires-python = ">= 3.11"

dependencies = [
    "uvicorn>=0.29.0",
    "fastapi>=0.110.0",
    "fastapi_async_sqlalchemy>=0.6.1",
    "sqlmodel>=0.0.16",
    "pydantic[email]>=2.6.4",
    "pydantic-settings>=2.2.1",
    "alembic>=1.13.1",
    "sqlalchemy_utils>=0.41.1",
    "websockets>=12.0",
    "python-dotenv>=1.0.1",
    "asyncpg>=0.29.0",
    "aiosqlite>=0.20.0",
    "asyncer>=0.0.5",
    "httpx>=0.27.0",
    "greenlet>=3.0.3",
    "aiohttp>=3.9.5",
    "cryptograpy>=0.0.0",
    "pyjwt>=2.8.0",
    "pyyaml>=6.0.1",
    "async_lru>=2.0.4",
    "httpx-ws>=0.5.2",
    "async-timeout>=4.0.3",
    "ruff>=0.3.2",
    "pytest>=8.1.1",
    "pytest-asyncio>=0.23.5.post1",
    "pytest-cov>=5.0.0",
    "mergedeep>=1.3.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "mpkernel @ file:///Users/boser/Dropbox/iot/mpkernel",
    "ipykernel>=6.29.3",
    "jupyter>=1.0.0",
    "jupyterlab>=4.1.5",
    "jupyter-book>=1.0.0",
    "pre-commit>=3.7.0",
    "ghp-import>=2.1.0",
    "pdoc>=14.4.0",
    "mpremote>=1.22.0",
    "esptool>=4.7.0",
    "pytest_mock>=3.14.0",
    "javascripthon>=0.12",
]

[tool.ruff]
line-length = 120
target-version = "py311"
exclude = ["alembic"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["earth/backend/app", "eventbus/eventbus"]
